Revert "Merge branch 'dev' into TWB-553-Create-Profile-Screen"
This reverts commit 8c4a50b88033397c05966c7c07ce1d393f6a4aa7, reversing
changes made to 9602b5fb3bd887a9e3377522f3eb81741eff7182.
