+++ TWB-Ecommerce.xcodeproj/xcuserdata/khurram.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -7,6 +7,22 @@
       <BreakpointProxy
          BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
          <BreakpointContent
+            uuid = "4FCF6C55-3B22-4F2D-99EA-0C9FDC14D414"
+            shouldBeEnabled = "No"
+            ignoreCount = "0"
+            continueAfterRunningActions = "No"
+            filePath = "TWB-Ecommerce/Features/Home/HomeView.swift"
+            startingColumnNumber = "9223372036854775807"
+            endingColumnNumber = "9223372036854775807"
+            startingLineNumber = "15"
+            endingLineNumber = "15"
+            landmarkName = "body"
+            landmarkType = "24">
+         </BreakpointContent>
+      </BreakpointProxy>
+      <BreakpointProxy
+         BreakpointExtensionID = "Xcode.Breakpoint.FileBreakpoint">
+         <BreakpointContent
             uuid = "82D0F0A7-256B-4E83-B1A7-8BB0739FDEBB"
             shouldBeEnabled = "Yes"
             ignoreCount = "0"
