--- TWB-Ecommerce/Features/Listing Screen/Views/Components/Filters/ColorsFilter/ColorOption.swift
@@ -1,16 +1,0 @@
-//
-//  ColorOption.swift
-//  TWB-Ecommerce
-//
-//  Created by Khurram Ansar on 24/09/2024.
-//
-
-import SwiftUI
-
-// Model for color options
-struct ColorOption: Identifiable {
-    let id = UUID()
-    let name: String
-    let color: Color
-}
-
