--- TWB-Ecommerce/Features/Listing Screen/Views/Components/Filters/FlowerTypeFilter/Flowers.swift
@@ -1,14 +1,0 @@
-//
-//  Flowers.swift
-//  TWB-Ecommerce
-//
-//  Created by Khurram Ansar on 24/09/2024.
-//
-
-import SwiftUI
-
-struct Flowers: Identifiable {
-    let id = UUID()
-    let name: String
-    let count: String
-}
