--- TWB-Ecommerce/Features/Listing Screen/Views/Components/Filters/OccasionsFilter/Occasion.swift
@@ -1,14 +1,0 @@
-//
-//  Untitled.swift
-//  TWB-Ecommerce
-//
-//  Created by Khurram Ansar on 24/09/2024.
-//
-
-import SwiftUI
-
-struct Occasion: Identifiable {
-    let id = UUID()
-    let name: String
-    let count: String
-}
