--- TWB-Ecommerce/Features/Listing Screen/Views/Components/Filters/PriceFilter/Prices.swift
@@ -1,14 +1,0 @@
-//
-//  Prices.swift
-//  TWB-Ecommerce
-//
-//  Created by Khurram Ansar on 24/09/2024.
-//
-
-import SwiftUI
-
-struct Prices: Identifiable {
-    let id = UUID()
-    let name: String
-    let count: String
-}
